test: reorganize and deduplicate unit + e2e tests
Phase A — Pure deduplication:
- Refactor core-regression-ui.spec.ts and core-regression.spec.ts to
  import helpers from e2e-navigation/e2e-api instead of inlining copies
- Add ensureInspectorLandsOnReportsHistory and openCreateFormPage helpers
  to e2e-navigation.ts (canonical, single source of truth)
- Merge screens/create-form.spec.ts (3 tests) into root create-form.spec.ts
  (6 tests → 9 tests); delete the screens copy
- Remove duplicate 'customer → not-allowed-page' tests from
  all-inventory.spec.ts and inventory-by-users.spec.ts; keep canonical
  RBAC matrix in navigation-and-rbac.spec.ts (expanded to cover
  /inventory-by-users route)
- Delete example.spec.ts (dead spec not wired to any nx target)
- Delete test-only maybeReverse describe block from reports-pdf-font.spec.ts
  (tested a locally-defined helper, not production code)
- Replace hardcoded Google OAuth client ID literal in
  google-sign-in.component.spec.ts with environment.googleClientId import

Phase B — Targeted improvements:
- Fix misleading jwt.service 'expired' test: generate a real past-exp
  RS256 token and assert 'Authentication token has expired' instead of
  using the literal string 'expired' which exercised the invalid-token
  branch, not the expiry branch
- Fix jwt.service.ts bug: TokenExpiredError extends JsonWebTokenError, so
  check the more specific type first (expiry branch was previously
  unreachable)
- Delete apps/backend-e2e/ project (Nx scaffold with one bogus test
  asserting GET / returns { message: 'Hello API' } which no real route
  serves; not referenced in any CI workflow)

Docs:
- Update README.md and package.json precommit script to remove
  backend-e2e references
