--- apps/backend-e2e/project.json
@@ -1,17 +1,0 @@
-{
-  "name": "backend-e2e",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "application",
-  "implicitDependencies": ["backend"],
-  "targets": {
-    "e2e": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{e2eProjectRoot}"],
-      "options": {
-        "jestConfig": "apps/backend-e2e/jest.config.ts",
-        "passWithNoTests": true
-      },
-      "dependsOn": ["backend:build"]
-    }
-  }
-}
