--- apps/backend-e2e/src/backend/backend.spec.ts
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-describe('GET /', () => {
-  it('should return a message', async () => {
-    const res = await axios.get(`/`);
-
-    expect(res.status).toBe(200);
-    expect(res.data).toEqual({ message: 'Hello API' });
-  });
-});
