--- apps/backend-e2e/src/support/test-setup.ts
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-module.exports = async function () {
-  // Configure axios for tests to use.
-  const host = process.env.HOST ?? 'localhost';
-  const port = process.env.PORT ?? '3000';
-  axios.defaults.baseURL = `http://${host}:${port}`;
-};
