+++ apps/frontend-e2e/src/screens/inventory-by-users.spec.ts
@@ -145,23 +145,4 @@ test.describe('inventory-by-users screen', () => {
     }
   });
 
-  test('customer cannot access inventory-by-users', async ({
-    page,
-    request,
-  }) => {
-    const token = await mintE2eJwt(request, {
-      backendBaseUrl,
-      e2eSecret,
-      userId: 'user-e2e-customer',
-      orgIdToRole: { [E2E_ORG_ID]: UserRole.Customer },
-    });
-
-    await bootstrapAuthenticatedSession(page, token, E2E_ORG_ID);
-    await ensureOrganizationIsSelected(page, E2E_ORG_ID);
-    await page.goto('/inventory-by-users');
-
-    await expect(
-      page.getByTestId('not-allowed-page')
-    ).toBeVisible({ timeout: 20000 });
-  });
 });
