+++ package.json
@@ -18,7 +18,7 @@
     "frontend:serve:e2e-local": "node scripts/write-runtime-config.js http://localhost:3000 && nx run frontend:serve",
     "e2e:local:test": "npm run e2e:local:prepare && npm run e2e:local:install-browsers && PLAYWRIGHT_HTML_OPEN=never E2E_SKIP_LOCAL_E2E_ENSURE=true npx nx run frontend-e2e:e2e-local-core",
     "e2e:deployed:test": "npm run e2e:local:install-browsers && PLAYWRIGHT_HTML_OPEN=never npx nx run frontend-e2e:e2e-deployed-core",
-    "precommit": "nx affected -t lint,test --uncommitted --nxBail --exclude=frontend-e2e,backend-e2e && npm run validate:translations",
+    "precommit": "nx affected -t lint,test --uncommitted --nxBail --exclude=frontend-e2e && npm run validate:translations",
     "prepare": "husky"
   },
   "private": true,
